Split event handeling from general miniwin features
