+++ Source/controls/game_controls.cpp
@@ -4,7 +4,7 @@
 
 #include "controls/controller.h"
 #include "controls/controller_motion.h"
-#include "miniwin/miniwin.h"
+#include "miniwin/misc_msg.h"
 #ifndef USE_SDL1
 #include "controls/devices/game_controller.h"
 #endif
