+++ Source/controls/plrctrls.cpp
@@ -11,7 +11,7 @@
 #include "automap.h"
 #include "control.h"
 #include "controls/controller_motion.h"
-#include "miniwin/miniwin.h"
+#include "miniwin/misc_msg.h"
 #ifndef USE_SDL1
 #include "controls/devices/game_controller.h"
 #endif
