+++ Source/inv.cpp
@@ -20,7 +20,7 @@
 #include "inv_iterators.hpp"
 #include "levels/town.h"
 #include "minitext.h"
-#include "miniwin/miniwin.h"
+#include "miniwin/misc_msg.h"
 #include "options.h"
 #include "panels/ui_panels.hpp"
 #include "plrmsg.h"
