+++ Source/levels/drlg_l4.cpp
@@ -5,10 +5,10 @@
  */
 #include "levels/drlg_l4.h"
 
-#include "miniwin/miniwin.h"
 #include "engine/load_file.hpp"
 #include "engine/random.hpp"
 #include "levels/gendung.h"
+#include "miniwin/miniwin.h"
 #include "monster.h"
 #include "multi.h"
 #include "objdat.h"
