+++ Source/movie.cpp
@@ -10,7 +10,7 @@
 #include "engine/demomode.h"
 #include "engine/sound.h"
 #include "hwcursor.hpp"
-#include "miniwin/miniwin.h"
+#include "miniwin/misc_msg.h"
 #include "storm/storm_svid.h"
 #include "utils/display.h"
