+++ Source/player.cpp
@@ -13,6 +13,7 @@
 #include "cursor.h"
 #include "dead.h"
 #include "miniwin/miniwin.h"
+#include "miniwin/misc_msg.h"
 #ifdef _DEBUG
 #include "debug.h"
 #endif
