+++ Source/utils/language.cpp
@@ -6,6 +6,7 @@
 #include <vector>
 
 #include "engine/assets.hpp"
+#include "miniwin/miniwin.h"
 #include "options.h"
 #include "utils/file_util.h"
 #include "utils/log.hpp"
